requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Settings + validation
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# Background jobs
rq==2.0.0
redis==5.2.1

# File parsing
pandas==2.2.3
openpyxl==3.1.5       # xlsx support for pandas
pdfplumber==0.11.4    # PDF extraction (stubbed in v1; wired for v2)

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18  # form/file uploads

# Utilities
python-dateutil==2.9.0
httpx==0.28.1         # async http client (also used in tests)

# Observability
structlog==24.4.0
